requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
instagram-private-api==1.6.0
Pillow==8.0.1
requests==2.25.0
urllib3==1.26.1
